requirements.txt
numpy>=1.24
gensim>=4.3
pyyaml>=6.0
requests>=2.31
